Changed console to .Net Core 1.0 and libraries to .Net Standard 1.6

+++ MpqTool/Foole.Mpq/MpqStream.cs
@@ -241,7 +241,7 @@ namespace Foole.Mpq
 
             int bytestocopy = Math.Min((int)(_currentData.Length - _position), count);
 
-            Array.Copy(_currentData, _position, buffer, offset, bytestocopy);
+            Array.Copy(_currentData, (int)_position, buffer, offset, bytestocopy);
 
             _position += bytestocopy;
             return bytestocopy;
